PY-105 Move tests.e2e.alpha.data to tests.e2e.data
Avoid importing alpha things from non-alpha tests
