+++ tests/e2e/conftest.py
@@ -20,7 +20,9 @@ from neptune_fetcher.util import (
     create_auth_api_client,
     get_config_and_token_urls,
 )
-from tests.e2e.alpha.data import (
+
+# TODO: un-alphafy
+from tests.e2e.data import (
     NOW,
     PATH,
     TEST_DATA,
